fix bug in flight factory class
 - update the flight model

 - update test to assert for fields from database query
