+++ api/models.py
@@ -7,5 +7,5 @@ class Flight(models.Model):
     origin = models.CharField(max_length=100)
     destination = models.CharField(max_length=100)
     departure_time = models.TimeField()
-    departure_date = models.DateField
+    departure_date = models.DateField(null=True)
     number = models.IntegerField()
